chore: upgrade Pallas to v0.30.1

+++ src/filters/legacy_v1/map.rs
@@ -408,7 +408,7 @@ impl EventWriter<'_> {
                 conway::RedeemerTag::Vote => "vote".to_string(),
                 conway::RedeemerTag::Propose => "propose".to_string(),
             },
-            ex_units_mem: redeemer.ex_units().mem,
+            ex_units_mem: redeemer.ex_units().mem as u32,
             ex_units_steps: redeemer.ex_units().steps,
             input_idx: redeemer.index(),
             plutus_data: redeemer.data().to_json(),
